+++ src/utils/errors/errorCode.error.ts
@@ -15,4 +15,5 @@ export class ErrorCode {
     public static readonly SendEmaillError = 'SendEmaillError';
     public static readonly InvalidInput = 'InvalidInput';
     public static readonly TooManyRequests = 'TooManyRequests';
+    public static readonly PayloadTooLarge = 'PayloadTooLarge';
   }